--- Alias/Controller/ViewController.swift
@@ -1,17 +1,0 @@
-//
-//  ViewController.swift
-//  Alias
-//
-//  Created by Maxim Vagin on 02.05.2022.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-    }
-
-}
